requirements.txt
ipykernel==5.1.4
jupyter==1.0.0
joblib==1.0.1
flask==1.1.1
google-resumable-media==0.4.1
google-cloud-storage==1.18.1
numpy==1.19.5
tensorflow==2.4.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vis==0.4.1